+++ assembly/src/assembler/linker/name_resolver.rs
@@ -2,7 +2,7 @@ use alloc::{borrow::Cow, collections::BTreeSet, vec::Vec};
 
 use super::{Linker, ModuleLink, PreLinkModule};
 use crate::{
-    RpoDigest, SourceSpan, Span, Spanned,
+    SourceSpan, Span, Spanned, Word,
     assembler::{GlobalProcedureIndex, LinkerError, ModuleIndex},
     ast::{
         Ident, InvocationTarget, InvokeKind, ProcedureName, QualifiedProcedureName,
@@ -58,7 +58,7 @@ pub enum ResolvedTarget {
     },
     /// We know the MAST root of the callee, but the procedure is not available, either in the
     /// procedure cache or in the module graph.
-    Phantom(RpoDigest),
+    Phantom(Word),
 }
 
 impl ResolvedTarget {